Separate tracing functions from output template factories
